refactor: component directories
