+++ frontend/src/components/applications/ApplicationInfoCard.vue
@@ -15,8 +15,8 @@ import {
 } from '@nethesis/vue-components'
 import { useApplicationDetail } from '@/queries/applications/applicationDetail'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
-import DataItem from '@/components/DataItem.vue'
-import NotesModal from '@/components/NotesModal.vue'
+import DataItem from '@/components/common/DataItem.vue'
+import NotesModal from '@/components/common/NotesModal.vue'
 import OrganizationIconAndLink from '@/components/organizations/OrganizationIconAndLink.vue'
 import { getDisplayName } from '@/lib/applications/applications'
 import ApplicationLogo from '@/components/applications/ApplicationLogo.vue'
