+++ frontend/src/components/applications/ApplicationSystemCard.vue
@@ -8,7 +8,7 @@ import { NeCard, NeHeading, NeSkeleton } from '@nethesis/vue-components'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { faServer } from '@fortawesome/free-solid-svg-icons'
 import { useApplicationDetail } from '@/queries/applications/applicationDetail'
-import DataItem from '@/components/DataItem.vue'
+import DataItem from '@/components/common/DataItem.vue'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 import SystemLogoAndLink from '../systems/SystemLogoAndLink.vue'
