+++ frontend/src/components/applications/ApplicationsTable.vue
@@ -40,7 +40,7 @@ import SetNotesDrawer from './SetNotesDrawer.vue'
 import { useApplicationFilters } from '@/queries/applications/applicationFilters'
 import { buildVersionFilterOptions } from '@/lib/applications/applicationFilters'
 import router from '@/router'
-import UpdatingSpinner from '@/components/UpdatingSpinner.vue'
+import UpdatingSpinner from '@/components/common/UpdatingSpinner.vue'
 import SystemDropdownFilter from '@/components/systems/SystemDropdownFilter.vue'
 import OrganizationDropdownFilter from '@/components/organizations/OrganizationDropdownFilter.vue'
 import SystemLogoAndLink from '../systems/SystemLogoAndLink.vue'
