+++ frontend/src/components/common/ClickToCopy.vue
