+++ frontend/src/components/common/CounterCard.vue
