+++ frontend/src/components/common/DeleteObjectModal.vue
