+++ frontend/src/components/common/ErrorModal.vue
