+++ frontend/src/components/common/NotesModal.vue
