+++ frontend/src/components/common/UpdatingSpinner.vue
