+++ frontend/src/components/distributors/DestroyDistributorModal.vue
@@ -13,7 +13,7 @@ import {
   type Distributor,
 } from '@/lib/organizations/distributors'
 import { useNotificationsStore } from '@/stores/notifications'
-import DeleteObjectModal from '../DeleteObjectModal.vue'
+import DeleteObjectModal from '../common/DeleteObjectModal.vue'
 
 const { visible = false, distributor = undefined } = defineProps<{
   visible: boolean
