+++ frontend/src/components/distributors/DistributorInfoCard.vue
@@ -15,9 +15,9 @@ import {
 import { useDistributorDetail } from '@/queries/organizations/distributorDetail'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { getOrganizationIcon } from '@/lib/organizations/organizations'
-import DataItem from '../DataItem.vue'
+import DataItem from '../common/DataItem.vue'
 import { ref } from 'vue'
-import NotesModal from '../NotesModal.vue'
+import NotesModal from '../common/NotesModal.vue'
 import { canManageDistributors } from '@/lib/permissions'
 import { faPenToSquare } from '@fortawesome/free-solid-svg-icons'
 import { useI18n } from 'vue-i18n'
