+++ frontend/src/components/shell/ImpersonationBadge.vue
