+++ frontend/src/components/shell/ImpersonationConsentBadge.vue
