+++ frontend/src/components/shell/ToastNotificationsArea.vue
@@ -6,7 +6,7 @@
 <script setup lang="ts">
 import { computed } from 'vue'
 import { useNotificationsStore } from '../../stores/notifications'
-import ErrorModal from '../ErrorModal.vue'
+import ErrorModal from '../common/ErrorModal.vue'
 import { type NeNotificationV2, NeToastNotificationV2 } from '@nethesis/vue-components'
 import { useI18n } from 'vue-i18n'
