+++ frontend/src/components/systems/SystemStatusCard.vue
@@ -23,7 +23,7 @@ import { useI18n } from 'vue-i18n'
 import { useSystemDetail } from '@/queries/systems/systemDetail'
 import { useSystemActiveAlerts } from '@/queries/systems/activeAlerts'
 import { useSystemBackups } from '@/queries/systems/backups'
-import DataItem from '../DataItem.vue'
+import DataItem from '../common/DataItem.vue'
 import { computed } from 'vue'
 import SystemStatusIcon from './SystemStatusIcon.vue'
 import type { Ns8Facts } from '@/lib/systems/ns8Facts'
