+++ frontend/src/views/ResellerDetailView.vue
@@ -9,7 +9,7 @@ import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { faArrowLeft, faServer, faBuilding } from '@fortawesome/free-solid-svg-icons'
 import { useResellerDetail } from '@/queries/organizations/resellerDetail'
 import ResellerInfoCard from '@/components/resellers/ResellerInfoCard.vue'
-import CounterCard from '@/components/CounterCard.vue'
+import CounterCard from '@/components/common/CounterCard.vue'
 import { useResellerStats } from '@/queries/organizations/resellerStats'
 import { faGridOne } from '@nethesis/nethesis-solid-svg-icons'
 import { useResellerSystems } from '@/queries/systems/resellerSystems'
